@@ -2,9 +2,13 @@
   "name": "mautic/grapes-js-builder-bundle",
   "description": "GrapesJS Builder with MJML support for Mautic",
   "type": "mautic-plugin",
-  "keywords": ["mautic","plugin","integration"],
-  "require": {
-    "php": "^7.2|^8.0"
+  "keywords": [
+    "mautic",
+    "plugin",
+    "integration"
+  ],
+  "extra": {
+    "install-directory-name": "GrapesJsBuilderBundle"
   },
   "require-dev": {
     "phpstan/phpstan": "^0.11.12",
@@ -26,5 +30,9 @@
     "csfixer": "vendor/bin/ecs check .",
     "fixcs": "vendor/bin/ecs check . --fix"
   },
-  "minimum-stability": "dev"
+  "minimum-stability": "dev",
+  "require": {
+    "php": ">=7.4.0 <8.1",
+    "mautic/core-lib": "^4.0"
+  }
 }
