@@ -49,6 +49,17 @@ class AppKernel extends Kernel
         defined('MAUTIC_ENV') or define('MAUTIC_ENV', $environment);
         defined('MAUTIC_VERSION') or define('MAUTIC_VERSION', $metadata->getVersion());
 
+        /**
+         * This is required for Doctrine's automatic database detection. When Mautic hasn't been
+         * installed yet, we don't have a database to connect to, causing automatic database platform
+         * detection to fail. We use the MAUTIC_DB_SERVER_VERSION constant to temporarily set a server_version
+         * if no database settings have been provided yet.
+         */
+        if (!defined('MAUTIC_DB_SERVER_VERSION')) {
+            $localConfigFile = ParameterLoader::getLocalConfigFile($this->getRootDir(), false);
+            define('MAUTIC_DB_SERVER_VERSION', file_exists($localConfigFile) ? null : '5.7');
+        }
+
         parent::__construct($environment, $debug);
     }
 
@@ -324,7 +335,6 @@ class AppKernel extends Kernel
      */
     public function getLocalConfigFile(): string
     {
-        /** @var $paths */
         $root = $this->getRootDir();
 
         return ParameterLoader::getLocalConfigFile($root);
