@@ -145,7 +145,7 @@ class EmailSubscriber implements EventSubscriberInterface
                 $stat->upRetryCount();
 
                 $retries = $stat->getRetryCount();
-                if (true || $retries > 3) {
+                if ($retries > 3) {
                     //tried too many times so just fail
                     $reason = $this->translator->trans('mautic.email.dnc.retries', [
                         '%subject%' => EmojiHelper::toShort($message->getSubject()),
