@@ -76,16 +76,16 @@ class FilterSelectorType extends AbstractType
         $builder->addEventListener(
             FormEvents::PRE_SET_DATA,
             function (FormEvent $event) use ($formModifier) {
-                $data = $event->getData();
-                $formModifier($event->getForm(), $data['column']);
+                $data = !empty($event->getData()) ? $event->getData()['column'] : null;
+                $formModifier($event->getForm(), $data);
             }
         );
 
         $builder->addEventListener(
             FormEvents::PRE_SUBMIT,
             function (FormEvent $event) use ($formModifier) {
-                $data = $event->getData();
-                $formModifier($event->getForm(), $data['column']);
+                $data = !empty($event->getData()) ? $event->getData()['column'] : null;
+                $formModifier($event->getForm(), $data);
             }
         );
 
